moved helper code setting up for refactor
